Remove redundant as c_int casts from comparisons and bitmask ops
Replace byte-to-c_int comparisons (u8 as c_int == 0xNN) with
direct byte comparisons, and remove as c_int from bitmask
operations (u8 as c_int & 0xNN). Clean up unused c_int imports.

+++ liblzma-rs/src/common/block_header_decoder.rs
@@ -1,5 +1,4 @@
 use crate::types::*;
-use core::ffi::c_int;
 extern "C" {
     fn lzma_vli_decode(
         vli: *mut lzma_vli,
@@ -58,11 +57,11 @@ pub unsafe extern "C" fn lzma_block_header_decode(
     if lzma_crc32(in_0, in_size, 0) != read32le(in_0.offset(in_size as isize)) {
         return LZMA_DATA_ERROR;
     }
-    if *in_0.offset(1) as c_int & 0x3c != 0 {
+    if *in_0.offset(1) & 0x3c != 0 {
         return LZMA_OPTIONS_ERROR;
     }
     let mut in_pos: size_t = 2;
-    if *in_0.offset(1) as c_int & 0x40 != 0 {
+    if *in_0.offset(1) & 0x40 != 0 {
         let ret_: lzma_ret = lzma_vli_decode(
             &raw mut (*block).compressed_size,
             core::ptr::null_mut(),
@@ -79,7 +78,7 @@ pub unsafe extern "C" fn lzma_block_header_decode(
     } else {
         (*block).compressed_size = LZMA_VLI_UNKNOWN;
     }
-    if *in_0.offset(1) as c_int & 0x80 != 0 {
+    if *in_0.offset(1) & 0x80 != 0 {
         let ret__0: lzma_ret = lzma_vli_decode(
             &raw mut (*block).uncompressed_size,
             core::ptr::null_mut(),
